Time: 1 ms (90.29%) | Memory: 42.8 MB (43.82%) - LeetSync

+++ 29-divide-two-integers/divide-two-integers.java
@@ -4,6 +4,6 @@ class Solution {
         int n=dividend/divisor;
         if(dividend<=-2147483648 && divisor==1) return -2147483648;
         if(dividend<=-2147483648 && divisor==-1) return 2147483647;
-        return (int) n;
+        return n;
     }
 }
